Updated how routes and handlers are organized
